refactor(dive-log): extract picker sheets from dive edit page
