+++ lib/features/dive_log/presentation/pages/dive_edit_page.dart
@@ -9,17 +9,14 @@ import 'package:submersion/core/constants/enums.dart';
 import 'package:submersion/features/marine_life/presentation/utils/species_category_icon.dart';
 import 'package:submersion/core/deco/altitude_calculator.dart';
 import 'package:submersion/core/services/location_service.dart';
-import 'package:submersion/core/text/fuzzy_match.dart';
 import 'package:submersion/core/utils/unit_formatter.dart';
 import 'package:submersion/features/buddies/domain/entities/buddy.dart';
 import 'package:submersion/features/buddies/presentation/providers/buddy_providers.dart';
 import 'package:submersion/features/buddies/presentation/widgets/buddy_picker.dart';
 import 'package:submersion/features/dive_sites/domain/entities/dive_site.dart';
 import 'package:submersion/features/dive_sites/presentation/providers/site_providers.dart';
-import 'package:submersion/features/dive_sites/presentation/widgets/similar_value_hint.dart';
 import 'package:submersion/features/equipment/domain/entities/equipment_item.dart';
 import 'package:submersion/features/equipment/domain/entities/equipment_set.dart';
-import 'package:submersion/features/equipment/presentation/providers/equipment_providers.dart';
 import 'package:submersion/features/equipment/presentation/providers/equipment_set_providers.dart';
 import 'package:submersion/features/marine_life/domain/entities/species.dart';
 import 'package:submersion/features/marine_life/presentation/providers/species_providers.dart';
@@ -40,12 +37,17 @@ import 'package:submersion/features/dive_log/domain/entities/dive_data_source.da
 import 'package:submersion/features/dive_log/presentation/providers/dive_providers.dart';
 import 'package:submersion/features/dive_log/presentation/providers/outlier_suggestion_provider.dart';
 import 'package:submersion/features/dive_log/presentation/widgets/custom_field_input_row.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/computer_source_sheet.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/edit_sighting_sheet.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/equipment_picker_sheet.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/equipment_set_picker_sheet.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/site_picker_sheet.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/pickers/species_picker_sheet.dart';
 import 'package:submersion/features/divers/presentation/providers/diver_providers.dart';
 import 'package:submersion/features/dive_log/presentation/widgets/ccr_settings_panel.dart';
 import 'package:submersion/features/dive_log/presentation/widgets/dive_mode_selector.dart';
 import 'package:submersion/features/dive_log/presentation/widgets/scr_settings_panel.dart';
 import 'package:submersion/features/dive_log/presentation/widgets/tank_editor.dart';
-import 'package:submersion/features/dive_log/presentation/utils/site_picker_search.dart';
 import 'package:submersion/features/tides/presentation/providers/tide_providers.dart';
 import 'package:submersion/features/weather/presentation/providers/weather_providers.dart';
 import 'package:submersion/features/courses/domain/entities/course.dart';
@@ -1146,7 +1148,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
         minChildSize: 0.5,
         maxChildSize: 0.95,
         expand: false,
-        builder: (sheetContext, scrollController) => _SitePickerSheet(
+        builder: (sheetContext, scrollController) => SitePickerSheet(
           scrollController: scrollController,
           selectedSiteId: _selectedSite?.id,
           currentLocation: _currentLocation,
@@ -1572,7 +1574,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
       final selected = await showModalBottomSheet<DiveDataSource>(
         context: context,
         builder: (context) =>
-            _ComputerSourceSelectionSheet(readings: originalReadings),
+            ComputerSourceSelectionSheet(readings: originalReadings),
       );
 
       if (selected == null || !mounted) return;
@@ -2200,7 +2202,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
         minChildSize: 0.5,
         maxChildSize: 0.95,
         expand: false,
-        builder: (context, scrollController) => _EquipmentPickerSheet(
+        builder: (context, scrollController) => EquipmentPickerSheet(
           scrollController: scrollController,
           selectedEquipmentIds: _selectedEquipment.map((e) => e.id).toSet(),
           onEquipmentSelected: (equipment) {
@@ -2226,7 +2228,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
         minChildSize: 0.5,
         maxChildSize: 0.95,
         expand: false,
-        builder: (context, scrollController) => _EquipmentSetPickerSheet(
+        builder: (context, scrollController) => EquipmentSetPickerSheet(
           scrollController: scrollController,
           onSetSelected: (set, items) {
             setState(() {
@@ -3283,7 +3285,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
         minChildSize: 0.5,
         maxChildSize: 0.95,
         expand: false,
-        builder: (context, scrollController) => _SpeciesPickerSheet(
+        builder: (context, scrollController) => SpeciesPickerSheet(
           scrollController: scrollController,
           onSpeciesSelected: (species, count, notes) {
             setState(() {
@@ -3311,7 +3313,7 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
     showModalBottomSheet(
       context: context,
       isScrollControlled: true,
-      builder: (context) => _EditSightingSheet(
+      builder: (context) => EditSightingSheet(
         sighting: sighting,
         onSave: (updatedSighting) {
           setState(() {
@@ -3786,1124 +3788,5 @@ class _DiveEditPageState extends ConsumerState<DiveEditPage> {
       case AltitudeWarningLevel.severe:
         return Colors.red;
     }
-  }
-}
-
-/// Site picker bottom sheet with nearby site suggestions
-class _SitePickerSheet extends ConsumerStatefulWidget {
-  final ScrollController scrollController;
-  final String? selectedSiteId;
-  final LocationResult? currentLocation;
-  final void Function(DiveSite) onSiteSelected;
-  final VoidCallback onCreateNewSite;
-
-  const _SitePickerSheet({
-    required this.scrollController,
-    required this.selectedSiteId,
-    this.currentLocation,
-    required this.onSiteSelected,
-    required this.onCreateNewSite,
-  });
-
-  @override
-  ConsumerState<_SitePickerSheet> createState() => _SitePickerSheetState();
-}
-
-class _SitePickerSheetState extends ConsumerState<_SitePickerSheet> {
-  final _searchController = TextEditingController();
-  String _searchQuery = '';
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  /// Calculate distance from current location to a site in km
-  double? _distanceToSite(DiveSite site) {
-    if (widget.currentLocation == null || site.location == null) return null;
-    final distanceMeters = LocationService.instance.distanceBetween(
-      widget.currentLocation!.latitude,
-      widget.currentLocation!.longitude,
-      site.location!.latitude,
-      site.location!.longitude,
-    );
-    return distanceMeters / 1000; // Convert to km
-  }
-
-  /// Format distance for display
-  String _formatDistance(double km) {
-    if (km < 1) {
-      return '${(km * 1000).round()} m away';
-    } else if (km < 10) {
-      return '${km.toStringAsFixed(1)} km away';
-    } else {
-      return '${km.round()} km away';
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final sitesAsync = ref.watch(sitesProvider);
-    final colorScheme = Theme.of(context).colorScheme;
-    final normalizedQuery = _searchQuery.trim().toLowerCase();
-
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    context.l10n.diveLog_sitePicker_title,
-                    style: Theme.of(context).textTheme.titleLarge,
-                  ),
-                  if (widget.currentLocation != null)
-                    Row(
-                      children: [
-                        Icon(
-                          Icons.my_location,
-                          size: 14,
-                          color: colorScheme.primary,
-                        ),
-                        const SizedBox(width: 4),
-                        Text(
-                          context.l10n.diveLog_sitePicker_sortedByDistance,
-                          style: Theme.of(context).textTheme.bodySmall
-                              ?.copyWith(color: colorScheme.primary),
-                        ),
-                      ],
-                    ),
-                ],
-              ),
-              TextButton.icon(
-                onPressed: widget.onCreateNewSite,
-                icon: const Icon(Icons.add),
-                label: Text(context.l10n.diveLog_sitePicker_newDiveSite),
-              ),
-            ],
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 0, 16, 12),
-          child: TextField(
-            controller: _searchController,
-            decoration: InputDecoration(
-              hintText: context.l10n.diveSites_list_search_placeholder,
-              prefixIcon: const Icon(Icons.search),
-              suffixIcon: normalizedQuery.isNotEmpty
-                  ? IconButton(
-                      icon: const Icon(Icons.clear),
-                      tooltip: context
-                          .l10n
-                          .diveLog_speciesPicker_tooltip_clearSearch,
-                      onPressed: () {
-                        _searchController.clear();
-                        setState(() => _searchQuery = '');
-                      },
-                    )
-                  : null,
-              border: const OutlineInputBorder(),
-            ),
-            onChanged: (value) => setState(() => _searchQuery = value),
-          ),
-        ),
-        if (_searchQuery.trim().isNotEmpty)
-          Builder(
-            builder: (context) {
-              final sites = sitesAsync.value ?? const <DiveSite>[];
-              final hidden = sites
-                  .where((s) => !siteMatchesPickerQuery(s, normalizedQuery))
-                  .toList();
-              final match = findSimilar(
-                _searchQuery,
-                hidden.map((s) => s.name),
-              );
-              if (match == null) return const SizedBox.shrink();
-              final site = hidden.firstWhere((s) => s.name == match);
-              return SimilarValueHint(
-                query: _searchQuery,
-                candidates: [match],
-                onAccept: (_) => widget.onSiteSelected(site),
-              );
-            },
-          ),
-        const Divider(height: 1),
-        Expanded(
-          child: sitesAsync.when(
-            data: (sites) {
-              if (sites.isEmpty) {
-                return Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(
-                        Icons.location_off,
-                        size: 48,
-                        color: colorScheme.onSurfaceVariant,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        context.l10n.diveLog_sitePicker_noSites,
-                        style: Theme.of(context).textTheme.titleMedium,
-                      ),
-                      const SizedBox(height: 8),
-                      TextButton.icon(
-                        onPressed: widget.onCreateNewSite,
-                        icon: const Icon(Icons.add),
-                        label: Text(
-                          context.l10n.diveLog_sitePicker_addDiveSite,
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              }
-
-              // Sort sites by distance if we have current location
-              List<_SiteWithDistance> sortedSites;
-              if (widget.currentLocation != null) {
-                sortedSites = sites.map((site) {
-                  return _SiteWithDistance(site, _distanceToSite(site));
-                }).toList();
-                // Sort: sites with distance first (by distance), then sites without GPS
-                sortedSites.sort((a, b) {
-                  if (a.distance == null && b.distance == null) return 0;
-                  if (a.distance == null) return 1;
-                  if (b.distance == null) return 1;
-                  return a.distance!.compareTo(b.distance!);
-                });
-              } else {
-                sortedSites = sites
-                    .map((site) => _SiteWithDistance(site, null))
-                    .toList();
-              }
-
-              final visibleSites = normalizedQuery.isEmpty
-                  ? sortedSites
-                  : sortedSites.where((siteWithDistance) {
-                      return siteMatchesPickerQuery(
-                        siteWithDistance.site,
-                        normalizedQuery,
-                      );
-                    }).toList();
-
-              if (visibleSites.isEmpty) {
-                return Center(
-                  child: Text(
-                    context.l10n.diveSites_list_search_noResults(
-                      _searchQuery.trim(),
-                    ),
-                    style: Theme.of(context).textTheme.titleMedium,
-                    textAlign: TextAlign.center,
-                  ),
-                );
-              }
-
-              return ListView.builder(
-                controller: widget.scrollController,
-                itemCount: visibleSites.length,
-                itemBuilder: (context, index) {
-                  final siteWithDist = visibleSites[index];
-                  final site = siteWithDist.site;
-                  final distance = siteWithDist.distance;
-                  final isSelected = site.id == widget.selectedSiteId;
-                  final isNearby =
-                      distance != null && distance < 50; // Within 50km
-
-                  return ListTile(
-                    leading: CircleAvatar(
-                      backgroundColor: isSelected
-                          ? colorScheme.primaryContainer
-                          : isNearby
-                          ? colorScheme.tertiaryContainer
-                          : colorScheme.surfaceContainerHighest,
-                      child: Icon(
-                        isNearby ? Icons.near_me : Icons.location_on,
-                        color: isSelected
-                            ? colorScheme.onPrimaryContainer
-                            : isNearby
-                            ? colorScheme.onTertiaryContainer
-                            : colorScheme.onSurfaceVariant,
-                      ),
-                    ),
-                    title: Text(site.name),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        if (site.locationString.isNotEmpty)
-                          Text(site.locationString),
-                        if (distance != null)
-                          Text(
-                            _formatDistance(distance),
-                            style: Theme.of(context).textTheme.bodySmall
-                                ?.copyWith(
-                                  color: isNearby
-                                      ? colorScheme.tertiary
-                                      : colorScheme.onSurfaceVariant,
-                                  fontWeight: isNearby ? FontWeight.w600 : null,
-                                ),
-                          ),
-                      ],
-                    ),
-                    trailing: isSelected
-                        ? Icon(Icons.check_circle, color: colorScheme.primary)
-                        : null,
-                    onTap: () => widget.onSiteSelected(site),
-                  );
-                },
-              );
-            },
-            loading: () => const Center(child: CircularProgressIndicator()),
-            error: (error, _) => Center(
-              child: Text(
-                context.l10n.diveLog_sitePicker_errorLoading(error.toString()),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-/// Helper class to hold a site with its calculated distance
-class _SiteWithDistance {
-  final DiveSite site;
-  final double? distance;
-
-  _SiteWithDistance(this.site, this.distance);
-}
-
-/// Species picker bottom sheet with search
-class _SpeciesPickerSheet extends ConsumerStatefulWidget {
-  final ScrollController scrollController;
-  final void Function(Species species, int count, String notes)
-  onSpeciesSelected;
-
-  const _SpeciesPickerSheet({
-    required this.scrollController,
-    required this.onSpeciesSelected,
-  });
-
-  @override
-  ConsumerState<_SpeciesPickerSheet> createState() =>
-      _SpeciesPickerSheetState();
-}
-
-class _SpeciesPickerSheetState extends ConsumerState<_SpeciesPickerSheet> {
-  final _searchController = TextEditingController();
-  String _searchQuery = '';
-  SpeciesCategory? _selectedCategory;
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final speciesAsync = _searchQuery.isEmpty && _selectedCategory == null
-        ? ref.watch(allSpeciesProvider)
-        : _selectedCategory != null
-        ? ref.watch(speciesByCategoryProvider(_selectedCategory!))
-        : ref.watch(speciesSearchProvider(_searchQuery));
-
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                context.l10n.diveLog_speciesPicker_title,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              IconButton(
-                icon: const Icon(Icons.close),
-                tooltip: context.l10n.common_action_close,
-                onPressed: () => Navigator.of(context).pop(),
-              ),
-            ],
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: TextField(
-            controller: _searchController,
-            decoration: InputDecoration(
-              hintText: context.l10n.diveLog_speciesPicker_searchHint,
-              prefixIcon: const Icon(Icons.search),
-              suffixIcon: _searchQuery.isNotEmpty
-                  ? IconButton(
-                      icon: const Icon(Icons.clear),
-                      tooltip: context
-                          .l10n
-                          .diveLog_speciesPicker_tooltip_clearSearch,
-                      onPressed: () {
-                        _searchController.clear();
-                        setState(() => _searchQuery = '');
-                      },
-                    )
-                  : null,
-              border: const OutlineInputBorder(),
-            ),
-            onChanged: (value) {
-              setState(() {
-                _searchQuery = value;
-                if (value.isNotEmpty) {
-                  _selectedCategory = null;
-                }
-              });
-            },
-          ),
-        ),
-        const SizedBox(height: 8),
-        SizedBox(
-          height: 40,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            children: [
-              _buildCategoryChip(null, 'All'),
-              ...SpeciesCategory.values.map(
-                (category) =>
-                    _buildCategoryChip(category, category.displayName),
-              ),
-            ],
-          ),
-        ),
-        const Divider(height: 16),
-        Expanded(
-          child: speciesAsync.when(
-            data: (speciesList) {
-              if (speciesList.isEmpty) {
-                return Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(
-                        Icons.water,
-                        size: 48,
-                        color: Theme.of(context).colorScheme.onSurfaceVariant,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        _searchQuery.isNotEmpty
-                            ? context.l10n.diveLog_speciesPicker_noResults
-                            : context.l10n.diveLog_speciesPicker_noSpecies,
-                        style: Theme.of(context).textTheme.titleMedium,
-                      ),
-                      if (_searchQuery.isNotEmpty) ...[
-                        const SizedBox(height: 8),
-                        FilledButton.tonal(
-                          onPressed: () => _addCustomSpecies(_searchQuery),
-                          child: Text(
-                            context.l10n.diveLog_speciesPicker_addNew(
-                              _searchQuery,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ],
-                  ),
-                );
-              }
-
-              return ListView.builder(
-                controller: widget.scrollController,
-                itemCount: speciesList.length,
-                itemBuilder: (context, index) {
-                  final species = speciesList[index];
-                  return ListTile(
-                    leading: CircleAvatar(
-                      backgroundColor: _getCategoryColor(species.category),
-                      child: Icon(
-                        iconForSpeciesCategory(species.category),
-                        color: Colors.white,
-                        size: 20,
-                      ),
-                    ),
-                    title: Text(species.commonName),
-                    subtitle: species.scientificName != null
-                        ? Text(
-                            species.scientificName!,
-                            style: const TextStyle(fontStyle: FontStyle.italic),
-                          )
-                        : null,
-                    trailing: Text(
-                      species.category.displayName,
-                      style: Theme.of(context).textTheme.bodySmall,
-                    ),
-                    onTap: () => _showSightingDetails(species),
-                  );
-                },
-              );
-            },
-            loading: () => const Center(child: CircularProgressIndicator()),
-            error: (error, _) =>
-                Center(child: Text('Error loading species: $error')),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildCategoryChip(SpeciesCategory? category, String label) {
-    final isSelected = _selectedCategory == category && _searchQuery.isEmpty;
-    return Padding(
-      padding: const EdgeInsetsDirectional.only(end: 8),
-      child: FilterChip(
-        label: Text(label),
-        selected: isSelected,
-        onSelected: (selected) {
-          setState(() {
-            _selectedCategory = selected ? category : null;
-            if (selected) {
-              _searchController.clear();
-              _searchQuery = '';
-            }
-          });
-        },
-      ),
-    );
-  }
-
-  Color _getCategoryColor(SpeciesCategory category) {
-    switch (category) {
-      case SpeciesCategory.fish:
-        return Colors.blue;
-      case SpeciesCategory.shark:
-        return Colors.grey.shade700;
-      case SpeciesCategory.ray:
-        return Colors.indigo;
-      case SpeciesCategory.mammal:
-        return Colors.brown;
-      case SpeciesCategory.turtle:
-        return Colors.green.shade700;
-      case SpeciesCategory.invertebrate:
-        return Colors.purple;
-      case SpeciesCategory.coral:
-        return Colors.pink;
-      case SpeciesCategory.plant:
-        return Colors.green;
-      case SpeciesCategory.other:
-        return Colors.grey;
-    }
-  }
-
-  void _showSightingDetails(Species species) {
-    int count = 1;
-    final notesController = TextEditingController();
-
-    showDialog(
-      context: context,
-      builder: (dialogContext) => StatefulBuilder(
-        builder: (context, setDialogState) => AlertDialog(
-          title: Text(species.commonName),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  IconButton(
-                    icon: const Icon(Icons.remove_circle_outline),
-                    tooltip: 'Decrease count',
-                    onPressed: count > 1
-                        ? () => setDialogState(() => count--)
-                        : null,
-                  ),
-                  Container(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 16,
-                      vertical: 8,
-                    ),
-                    decoration: BoxDecoration(
-                      border: Border.all(
-                        color: Theme.of(context).colorScheme.outline,
-                      ),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: Text(
-                      '$count',
-                      style: Theme.of(context).textTheme.titleLarge,
-                    ),
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.add_circle_outline),
-                    tooltip: 'Increase count',
-                    onPressed: () => setDialogState(() => count++),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              TextField(
-                controller: notesController,
-                decoration: InputDecoration(
-                  labelText: context.l10n.diveLog_sighting_notesOptional,
-                  hintText: context.l10n.diveLog_sighting_notesHint,
-                  border: const OutlineInputBorder(),
-                ),
-                maxLines: 2,
-              ),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(dialogContext).pop(),
-              child: Text(context.l10n.diveLog_sighting_cancel),
-            ),
-            FilledButton(
-              onPressed: () {
-                Navigator.of(dialogContext).pop();
-                widget.onSpeciesSelected(species, count, notesController.text);
-              },
-              child: Text(context.l10n.diveLog_sighting_add),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _addCustomSpecies(String name) async {
-    final repository = ref.read(speciesRepositoryProvider);
-    final species = await repository.getOrCreateSpecies(
-      commonName: name,
-      category: SpeciesCategory.other,
-    );
-    if (mounted) {
-      _showSightingDetails(species);
-    }
-  }
-}
-
-/// Edit sighting sheet
-class _EditSightingSheet extends StatefulWidget {
-  final Sighting sighting;
-  final void Function(Sighting) onSave;
-  final VoidCallback onDelete;
-
-  const _EditSightingSheet({
-    required this.sighting,
-    required this.onSave,
-    required this.onDelete,
-  });
-
-  @override
-  State<_EditSightingSheet> createState() => _EditSightingSheetState();
-}
-
-class _EditSightingSheetState extends State<_EditSightingSheet> {
-  late int _count;
-  late TextEditingController _notesController;
-
-  @override
-  void initState() {
-    super.initState();
-    _count = widget.sighting.count;
-    _notesController = TextEditingController(text: widget.sighting.notes);
-  }
-
-  @override
-  void dispose() {
-    _notesController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(
-        left: 16,
-        right: 16,
-        top: 16,
-        bottom: MediaQuery.of(context).viewInsets.bottom + 16,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                widget.sighting.speciesName,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              IconButton(
-                icon: const Icon(Icons.delete_outline, color: Colors.red),
-                tooltip: context.l10n.diveLog_editSighting_remove,
-                onPressed: () {
-                  showDialog(
-                    context: context,
-                    builder: (dialogContext) => AlertDialog(
-                      title: Text(
-                        context.l10n.diveLog_editSighting_removeTitle,
-                      ),
-                      content: Text(
-                        context.l10n.diveLog_editSighting_removeConfirm(
-                          widget.sighting.speciesName,
-                        ),
-                      ),
-                      actions: [
-                        TextButton(
-                          onPressed: () => Navigator.of(dialogContext).pop(),
-                          child: Text(context.l10n.common_action_cancel),
-                        ),
-                        FilledButton(
-                          onPressed: () {
-                            Navigator.of(dialogContext).pop();
-                            widget.onDelete();
-                          },
-                          style: FilledButton.styleFrom(
-                            backgroundColor: Theme.of(
-                              context,
-                            ).colorScheme.error,
-                          ),
-                          child: Text(context.l10n.diveLog_editSighting_remove),
-                        ),
-                      ],
-                    ),
-                  );
-                },
-              ),
-            ],
-          ),
-          const SizedBox(height: 24),
-          Text(
-            context.l10n.diveLog_editSighting_count,
-            style: Theme.of(context).textTheme.labelLarge,
-          ),
-          const SizedBox(height: 8),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              IconButton.filled(
-                icon: const Icon(Icons.remove),
-                onPressed: _count > 1 ? () => setState(() => _count--) : null,
-              ),
-              Container(
-                margin: const EdgeInsets.symmetric(horizontal: 24),
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 24,
-                  vertical: 12,
-                ),
-                decoration: BoxDecoration(
-                  border: Border.all(
-                    color: Theme.of(context).colorScheme.outline,
-                  ),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: Text(
-                  '$_count',
-                  style: Theme.of(context).textTheme.headlineSmall,
-                ),
-              ),
-              IconButton.filled(
-                icon: const Icon(Icons.add),
-                onPressed: () => setState(() => _count++),
-              ),
-            ],
-          ),
-          const SizedBox(height: 24),
-          TextField(
-            controller: _notesController,
-            decoration: InputDecoration(
-              labelText: context.l10n.diveLog_editSighting_notes,
-              hintText: context.l10n.diveLog_editSighting_notesHint,
-              border: const OutlineInputBorder(),
-            ),
-            maxLines: 3,
-          ),
-          const SizedBox(height: 24),
-          FilledButton(
-            onPressed: () {
-              widget.onSave(
-                widget.sighting.copyWith(
-                  count: _count,
-                  notes: _notesController.text,
-                ),
-              );
-            },
-            child: Text(context.l10n.diveLog_editSighting_save),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// Equipment picker bottom sheet
-class _EquipmentPickerSheet extends ConsumerWidget {
-  final ScrollController scrollController;
-  final Set<String> selectedEquipmentIds;
-  final void Function(EquipmentItem) onEquipmentSelected;
-
-  const _EquipmentPickerSheet({
-    required this.scrollController,
-    required this.selectedEquipmentIds,
-    required this.onEquipmentSelected,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final equipmentAsync = ref.watch(allEquipmentProvider);
-
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                context.l10n.diveLog_equipmentPicker_title,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              IconButton(
-                icon: const Icon(Icons.close),
-                tooltip: context.l10n.common_action_close,
-                onPressed: () => Navigator.of(context).pop(),
-              ),
-            ],
-          ),
-        ),
-        const Divider(height: 1),
-        Expanded(
-          child: equipmentAsync.when(
-            data: (equipmentList) {
-              // Filter out already selected equipment
-              final available = equipmentList
-                  .where((e) => !selectedEquipmentIds.contains(e.id))
-                  .toList();
-
-              if (available.isEmpty) {
-                return Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(
-                        Icons.inventory_2_outlined,
-                        size: 48,
-                        color: Theme.of(context).colorScheme.onSurfaceVariant,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        equipmentList.isEmpty
-                            ? context.l10n.diveLog_equipmentPicker_noEquipment
-                            : context.l10n.diveLog_equipmentPicker_allSelected,
-                        style: Theme.of(context).textTheme.titleMedium,
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        equipmentList.isEmpty
-                            ? context.l10n.diveLog_equipmentPicker_addFromTab
-                            : context.l10n.diveLog_equipmentPicker_removeToAdd,
-                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                          color: Theme.of(context).colorScheme.onSurfaceVariant,
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              }
-
-              return ListView.builder(
-                controller: scrollController,
-                itemCount: available.length,
-                itemBuilder: (context, index) {
-                  final equipment = available[index];
-                  return ListTile(
-                    leading: CircleAvatar(
-                      backgroundColor: Theme.of(
-                        context,
-                      ).colorScheme.surfaceContainerHighest,
-                      child: Icon(
-                        _getEquipmentIcon(equipment.type),
-                        color: Theme.of(context).colorScheme.onSurfaceVariant,
-                      ),
-                    ),
-                    title: Text(equipment.name),
-                    subtitle: Text(equipment.type.displayName),
-                    onTap: () => onEquipmentSelected(equipment),
-                  );
-                },
-              );
-            },
-            loading: () => const Center(child: CircularProgressIndicator()),
-            error: (error, _) => Center(
-              child: Text(
-                context.l10n.diveLog_equipmentPicker_errorLoading(
-                  error.toString(),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  IconData _getEquipmentIcon(EquipmentType type) {
-    switch (type) {
-      case EquipmentType.regulator:
-        return Icons.air;
-      case EquipmentType.bcd:
-        return Icons.checkroom;
-      case EquipmentType.wetsuit:
-      case EquipmentType.drysuit:
-        return Icons.dry_cleaning;
-      case EquipmentType.mask:
-        return Icons.visibility;
-      case EquipmentType.fins:
-        return Icons.water;
-      case EquipmentType.boots:
-        return Icons.hiking;
-      case EquipmentType.gloves:
-        return Icons.pan_tool;
-      case EquipmentType.hood:
-        return Icons.face;
-      case EquipmentType.tank:
-        return MdiIcons.divingScubaTank;
-      case EquipmentType.weights:
-        return Icons.fitness_center;
-      case EquipmentType.computer:
-        return Icons.watch;
-      case EquipmentType.light:
-        return Icons.flashlight_on;
-      case EquipmentType.camera:
-        return Icons.camera_alt;
-      case EquipmentType.knife:
-        return Icons.content_cut;
-      case EquipmentType.smb:
-        return Icons.flag;
-      case EquipmentType.reel:
-        return Icons.all_inclusive;
-      case EquipmentType.other:
-        return Icons.build;
-    }
-  }
-}
-
-/// Equipment set picker bottom sheet
-class _EquipmentSetPickerSheet extends ConsumerWidget {
-  final ScrollController scrollController;
-  final void Function(EquipmentSet set, List<EquipmentItem> items)
-  onSetSelected;
-
-  const _EquipmentSetPickerSheet({
-    required this.scrollController,
-    required this.onSetSelected,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final setsAsync = ref.watch(equipmentSetsProvider);
-
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                context.l10n.diveLog_equipmentSetPicker_title,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              IconButton(
-                icon: const Icon(Icons.close),
-                tooltip: context.l10n.common_action_close,
-                onPressed: () => Navigator.of(context).pop(),
-              ),
-            ],
-          ),
-        ),
-        const Divider(height: 1),
-        Expanded(
-          child: setsAsync.when(
-            data: (sets) {
-              if (sets.isEmpty) {
-                return Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(
-                        Icons.folder_special_outlined,
-                        size: 48,
-                        color: Theme.of(context).colorScheme.onSurfaceVariant,
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        context.l10n.diveLog_equipmentSetPicker_noSets,
-                        style: Theme.of(context).textTheme.titleMedium,
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        context.l10n.diveLog_equipmentSetPicker_createHint,
-                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                          color: Theme.of(context).colorScheme.onSurfaceVariant,
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              }
-
-              return ListView.builder(
-                controller: scrollController,
-                itemCount: sets.length,
-                itemBuilder: (context, index) {
-                  final set = sets[index];
-                  return _EquipmentSetTile(
-                    set: set,
-                    onTap: (items) => onSetSelected(set, items),
-                  );
-                },
-              );
-            },
-            loading: () => const Center(child: CircularProgressIndicator()),
-            error: (error, _) => Center(
-              child: Text(
-                context.l10n.diveLog_equipmentSetPicker_errorLoading(
-                  error.toString(),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-/// Individual equipment set tile that loads its items
-class _EquipmentSetTile extends ConsumerWidget {
-  final EquipmentSet set;
-  final void Function(List<EquipmentItem> items) onTap;
-
-  const _EquipmentSetTile({required this.set, required this.onTap});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final setWithItemsAsync = ref.watch(equipmentSetWithItemsProvider(set.id));
-
-    return setWithItemsAsync.when(
-      data: (setWithItems) {
-        if (setWithItems == null) {
-          return const SizedBox.shrink();
-        }
-        final items = setWithItems.items ?? [];
-        return ListTile(
-          leading: CircleAvatar(
-            backgroundColor: Theme.of(context).colorScheme.primaryContainer,
-            child: Icon(
-              Icons.folder_special,
-              color: Theme.of(context).colorScheme.onPrimaryContainer,
-            ),
-          ),
-          title: Text(set.name),
-          subtitle: Text(
-            items.isEmpty
-                ? context.l10n.diveLog_equipmentSetPicker_emptySet
-                : '${items.length} item${items.length == 1 ? '' : 's'}: ${items.map((e) => e.name).join(', ')}',
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-          ),
-          trailing: const Icon(Icons.chevron_right),
-          onTap: items.isEmpty ? null : () => onTap(items),
-        );
-      },
-      loading: () => ListTile(
-        leading: CircleAvatar(
-          backgroundColor: Theme.of(
-            context,
-          ).colorScheme.surfaceContainerHighest,
-          child: const SizedBox(
-            width: 20,
-            height: 20,
-            child: CircularProgressIndicator(strokeWidth: 2),
-          ),
-        ),
-        title: Text(set.name),
-        subtitle: Text(context.l10n.diveLog_equipmentSetPicker_loading),
-      ),
-      error: (_, _) => ListTile(
-        leading: CircleAvatar(
-          backgroundColor: Theme.of(context).colorScheme.errorContainer,
-          child: Icon(
-            Icons.error_outline,
-            color: Theme.of(context).colorScheme.onErrorContainer,
-          ),
-        ),
-        title: Text(set.name),
-        subtitle: Text(context.l10n.diveLog_equipmentSetPicker_errorItems),
-      ),
-    );
-  }
-}
-
-/// Bottom sheet that lets the user choose which computer's profile to use
-/// as the starting point when opening the profile editor on a multi-computer
-/// dive.
-class _ComputerSourceSelectionSheet extends StatelessWidget {
-  final List<DiveDataSource> readings;
-
-  const _ComputerSourceSelectionSheet({required this.readings});
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final textTheme = Theme.of(context).textTheme;
-
-    return SafeArea(
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(16, 20, 16, 8),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text('Choose starting profile', style: textTheme.titleMedium),
-            const SizedBox(height: 4),
-            Text(
-              'Select which computer\'s profile to edit from.',
-              style: textTheme.bodySmall?.copyWith(
-                color: colorScheme.onSurfaceVariant,
-              ),
-            ),
-            const SizedBox(height: 12),
-            ...readings.map(
-              (reading) => ListTile(
-                leading: Icon(
-                  Icons.computer,
-                  color: reading.isPrimary
-                      ? colorScheme.primary
-                      : colorScheme.onSurfaceVariant,
-                ),
-                title: Text(reading.displayName),
-                subtitle: reading.isPrimary
-                    ? Text(
-                        'Primary',
-                        style: textTheme.bodySmall?.copyWith(
-                          color: colorScheme.primary,
-                        ),
-                      )
-                    : null,
-                onTap: () => Navigator.of(context).pop(reading),
-              ),
-            ),
-            const SizedBox(height: 8),
-          ],
-        ),
-      ),
-    );
   }
 }
